# Conflicts:
#	app/build.gradle
#	build.gradle

+++ app/src/main/java/com/xiaoe/shop/wxb/base/BaseFragment.java
@@ -12,7 +12,6 @@ import android.view.Gravity;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.webkit.WebChromeClient;
 import android.webkit.WebSettings;
 import android.webkit.WebView;
 import android.widget.PopupWindow;
@@ -27,7 +26,7 @@ import com.xiaoe.common.db.SQLiteUtil;
 import com.xiaoe.common.entitys.LoginUser;
 import com.xiaoe.common.utils.Dp2Px2SpUtil;
 import com.xiaoe.network.NetworkCodes;
-import com.xiaoe.network.NetworkStateResult;
+import com.xiaoe.common.entitys.NetworkStateResult;
 import com.xiaoe.network.network_interface.INetworkResponse;
 import com.xiaoe.network.requests.IRequest;
 import com.xiaoe.shop.wxb.R;
@@ -183,24 +182,29 @@ public class BaseFragment extends Fragment implements INetworkResponse, OnCustom
             public void run() {
                 JSONObject jsonObject = (JSONObject) entity;
                 if (success && entity != null) {
-                    if(jsonObject.getIntValue("code") == NetworkCodes.CODE_NOT_LOAING){
-                        if(!dialog.isShowing() && !((XiaoeActivity) getActivity()).getDialog().isShowing()){
-                            dialog.setCancelable(false);
-                            dialog.setHideCancelButton(true);
-                            dialog.getTitleView().setGravity(Gravity.START);
-                            dialog.getTitleView().setPadding(Dp2Px2SpUtil.dp2px(getContext(), 22), 0, Dp2Px2SpUtil.dp2px(getContext(), 22), 0 );
-                            dialog.getTitleView().setTextSize(TypedValue.COMPLEX_UNIT_SP, 14);
-                            dialog.setMessageVisibility(View.GONE);
-                            dialog.setCancelable(false);
-                            dialog.setHideCancelButton(true);
-                            dialog.setTitle(getString(R.string.login_invalid));
-                            dialog.setConfirmText(getString(R.string.btn_again_login));
-                            dialog.showDialog(DIALOG_TAG_LOADING);
-                            // 往回传 null 关闭加载中
-                            onMainThreadResponse(null, false, entity);
+                    try {
+                        if (jsonObject.getInteger("code") == NetworkCodes.CODE_NOT_LOAING) {
+                            if (!dialog.isShowing() && !((XiaoeActivity) getActivity()).getDialog().isShowing()) {
+                                dialog.setCancelable(false);
+                                dialog.setHideCancelButton(true);
+                                dialog.getTitleView().setGravity(Gravity.START);
+                                dialog.getTitleView().setPadding(Dp2Px2SpUtil.dp2px(getContext(), 22), 0, Dp2Px2SpUtil.dp2px(getContext(), 22), 0);
+                                dialog.getTitleView().setTextSize(TypedValue.COMPLEX_UNIT_SP, 14);
+                                dialog.setMessageVisibility(View.GONE);
+                                dialog.setCancelable(false);
+                                dialog.setHideCancelButton(true);
+                                dialog.setTitle(getString(R.string.login_invalid));
+                                dialog.setConfirmText(getString(R.string.btn_again_login));
+                                dialog.showDialog(DIALOG_TAG_LOADING);
+                                // 往回传 null 关闭加载中
+                                doResponseSuccess(null, false, entity);
+                            }
+                        } else {
+                            doResponseSuccess(iRequest, true, entity);
                         }
-                    }else{
-                        onMainThreadResponse(iRequest, true, entity);
+                    } catch (Exception e) {
+                        e.printStackTrace();
+                        ToastUtils.show(getContext(), getString(R.string.service_error_text));
                     }
                 } else {
                     if (jsonObject != null) {
@@ -210,10 +214,20 @@ public class BaseFragment extends Fragment implements INetworkResponse, OnCustom
                             ToastUtils.show(getContext(), getString(R.string.network_error_text));
                         }
                     }
-                    onMainThreadResponse(iRequest, false, entity);
+                    doResponseSuccess(iRequest, false, entity);
                 }
             }
         });
+    }
+
+    private void doResponseSuccess(IRequest iRequest, boolean success, Object entity) {
+        //防止由于请求后，处理返回数据崩溃的异常
+        try {
+            onMainThreadResponse(iRequest, success, entity);
+        } catch (Exception e) {
+            e.printStackTrace();
+            ToastUtils.show(getContext(), getString(R.string.service_error_text));
+        }
     }
 
     @Override
@@ -291,9 +305,13 @@ public class BaseFragment extends Fragment implements INetworkResponse, OnCustom
     public void onClickConfirm(View view, int tag) {
         if(tag == DIALOG_TAG_LOADING){
             SQLiteUtil sqLiteUtil = SQLiteUtil.init(getActivity(), new LoginSQLiteCallback());
-            sqLiteUtil.execSQL("delete from " + LoginSQLiteCallback.TABLE_NAME_USER);
+            if (sqLiteUtil.tabIsExist(LoginSQLiteCallback.TABLE_NAME_USER)) {
+                sqLiteUtil.execSQL("delete from " + LoginSQLiteCallback.TABLE_NAME_USER);
+            }
             SQLiteUtil lrUtil = SQLiteUtil.init(getActivity(), new LrSQLiteCallback());
-            lrUtil.execSQL("delete from " + LrSQLiteCallback.TABLE_NAME_LR);
+            if (lrUtil.tabIsExist(LrSQLiteCallback.TABLE_NAME_LR)) {
+                lrUtil.execSQL("delete from " + LrSQLiteCallback.TABLE_NAME_LR);
+            }
             CommonUserInfo.getInstance().clearUserInfo();
             CommonUserInfo.getInstance().clearLoginUserInfo();
             CommonUserInfo.setApiToken("");
